refactor(dashboard): I1 — move Step-10 auto-escalate classifier → codec_chat_pipeline (SR-60)
codec_dashboard.py: 1,310 → 1,201 LOC (-109).
Cumulative since pre-B6 baseline: 3,912 → 1,201 (-69.3%).

The Phase 3 Step 10 auto-escalation classifier cluster (9 names:
_AUTO_ESCALATE_SYSTEM_PROMPT, _qwen_chat_classify, _classify_chat_message,
_AUTOESCALATE_SILENCE_LOCK, _autoescalate_silence_set,
ESCALATE_CHECKPOINTS_THRESHOLD, silence_session_autoescalate,
_reset_autoescalate_silence_for_test, _should_escalate_to_project) moves
verbatim to codec_chat_pipeline.py — its natural home alongside _StepBudget /
_is_conversational. The cluster is latent (no production caller yet; the
chat→project "Promote?" prompt is deferred to Phase 3.5) but test-covered.

codec_dashboard re-exports all 9 names identity-equal for back-compat.

Test surface:
  - test_chat_escalation: monkeypatch target swapped codec_dashboard →
    codec_chat_pipeline. The functions call each other through the pipeline
    module namespace, so patching the in-module chain (_qwen_chat_classify →
    _classify_chat_message → _should_escalate_to_project) must target where
    they're DEFINED, not the re-export. All 8 monkeypatch tests updated.
  - test_dashboard_llm: codec_llm.call count >=2 → >=1 (classifier moved;
    only /api/command's Flash fallback remains in codec_dashboard).
  - new TestI1EscalationExtraction (3 pins): cluster-in-pipeline, 9-name
    identity-equal re-export, dashboard-no-longer-defines.
  - drive-by ruff F401: dropped now-unused `import threading` from codec_dashboard
    (the only consumer was _AUTOESCALATE_SILENCE_LOCK, which moved).

codec_chat_pipeline gains `import threading` + `import codec_llm` for the
moved code. test_dashboard_llm's 3 _qwen_chat_classify behavior tests are
unaffected (they patch codec_llm.call at source, which the re-export honors).

Full suite: 2,055 passed / 77 skipped (was 2,052 in H1). +3 net.
